feat[ia5.2]: implemented additional field for action & benefits exchange
