+++ HealthcareSystem/IntegrationAdapters/Views/Pharmacy/ApiRegister.cshtml
@@ -12,15 +12,15 @@
     <hr />
     @Html.ValidationSummary(true, "", new { @class = "text-danger" })
 
-    <div class="form-group" >
+    <div class="form-group">
         @if (TempData["Success"] != null)
-            {
-                <p class="alert alert-success" id="successMessage">@TempData["Success"]</p>
-            }
+        {
+            <p class="alert alert-success" id="successMessage">@TempData["Success"]</p>
+        }
     </div>
 
     <div class="form-group">
-        @Html.LabelFor(model => model.Name, htmlAttributes: new { @class = "control-label col-md-2" })
+        <label class="control-label col-md-2">Name</label>
         <div class="col-md-10">
             @Html.TextBoxFor(model => model.Name, new { @class = "form-control" })
             @Html.ValidationMessageFor(model => model.Name, "", new { @class = "text-danger" })
@@ -28,7 +28,7 @@
     </div>
 
     <div class="form-group">
-        @Html.LabelFor(model => model.ApiKey, htmlAttributes: new { @class = "control-label col-md-2" })
+        <label class="control-label col-md-2">API Key</label>
         <div class="col-md-10">
             @Html.TextBoxFor(model => model.ApiKey, new { @class = "form-control" })
             @Html.ValidationMessageFor(model => model.ApiKey, "", new { @class = "text-danger" })
@@ -36,7 +36,7 @@
     </div>
 
     <div class="form-group">
-        @Html.LabelFor(model => model.Url, htmlAttributes: new { @class = "control-label col-md-2" })
+        <label class="control-label col-md-2">Url</label>
         <div class="col-md-10">
             @Html.TextBoxFor(model => model.Url, new { @class = "form-control" })
             @Html.ValidationMessageFor(model => model.Url, "", new { @class = "text-danger" })
@@ -44,10 +44,17 @@
     </div>
 
     <div class="form-group">
-        <div class="col-md-10"></div>
-            <input type="submit" value="Submit" class="btn-blue-grey" />
+        <label class="control-label col-md-5">Actions & Benefits Exchange Name</label>
+        <div class="col-md-10">
+            @Html.TextBoxFor(model => model.ActionsBenefitsExchangeName, new { @class = "form-control" })
+        </div>
     </div>
-    
+
+    <div class="form-group">
+        <div class="col-md-10"></div>
+        <input type="submit" value="Submit" class="btn-blue-grey" />
+    </div>
+
 
 </div>
 }
