+++ HealthcareSystem/IntegrationAdapters/Views/Shared/_Layout.cshtml
@@ -23,7 +23,7 @@
                             <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Index">Home</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="ApiRegister">Register API</a>
+                            <a class="nav-link text-dark" asp-area="" asp-controller="Pharmacy" asp-action="ApiRegister">Register API</a>
                         </li>
                         <li class="nav-item">
                             <a class="nav-link text-dark" asp-area="" asp-controller="ActionBenefit" asp-action="Index">Actions&Benefits</a>
